+++ components/Header.vue
@@ -2,7 +2,7 @@
   <div class="card mb-5 border border-2 border-dark bg-warning">
     <div class="text-center py-1">
       <span class="fs-1 lexend-mega">e-PKL<span class="text-info dot-stroke">.</span></span>
-      <span class="fw-bold smallest ">peserta</span>
+      <span class="fw-bold smallest ">siswa</span>
     </div>
   </div>
 </template>