+++ layouts/login.vue
@@ -22,7 +22,8 @@
   src: url('~/assets/font/Metropolis-Regular.otf');
 }
 html, body {
-  background: #fff url('~/assets/img/bg.webp') no-repeat center bottom fixed;
+  /* background: #fff url('~/assets/img/bg.webp') no-repeat center bottom fixed; */
+  background-color: #fff;
   background-size: cover;
   font-family: "Quicksand", sans-serif;
 }