+++ pages/elemen.vue
@@ -10,12 +10,14 @@
           <li><code>Lain-lain</code> adalah opsi di luar konteks elemen</li>
         </ul>
       </div>
+
       <loading v-if="isLoading" />
-      <div v-if="elements" class="accordion accordion-flush border border-2 border-dark shadow-lg" id="accordion-parent">
-        <div v-for="(elemen,i) in elements" :key="elemen.id" class="accordion-item border-bottom border-2 border-dark">
+
+      <div v-if="elements" class="accordion accordion-flush" id="accordion-parent">
+        <div v-for="(elemen,i) in elements" :key="elemen.id" class="accordion-item border-bottom border-1 border-grey">
           <div class="accordion-header">
             <button class="accordion-button collapsed fw-bold" type="button" data-bs-toggle="collapse" :data-bs-target="`#flush-${elemen.id}`" aria-expanded="false" :aria-controls="`flush-${elemen.id}`">
-              <span class="badge text-dark border border-2 border-dark me-2">{{ i+1 }}</span> {{ elemen.elemen }}
+              <span class="badge text-muted border border-2 border-grey me-2">{{ i+1 }}</span> {{ elemen.elemen }}
             </button>
           </div>
           <div :id="`flush-${elemen.id}`" class="accordion-collapse collapse" data-bs-parent="#accordion-parent">
@@ -28,6 +30,7 @@
           </div>
         </div>
       </div>
+
     </div>
   </div>
 </template>